generate_demographic_tables

generate_demographic_tables This function creates demographic summary tables and calculates statistical tests between groups.

Usage

generate_demographic_tables(
  input_data = mtcars,
  parameter_list = c("mpg", "disp", "qsec"),
  group_column = "am",
  completed_table_statistic = "median",
  freeform_brackets = "Q1_Q3",
  groupwise_test = "wilcox",
  multiple_comparison_correction = "BH",
  comparison_list = NULL,
  multiple_comparison_correction_brackets = FALSE,
  groupwise_95CI = FALSE,
  sex_column = "vs",
  female_value = 0,
  male_value = 1,
  completed_table_sex_statistic = "n_m_f",
  group_n = TRUE,
  format_value = 2,
  p_format_value = 3,
  mean_if_normal = FALSE,
  ttest_if_normal = FALSE,
  ordinal_columns = NULL
)

Arguments

input_data

data.frame containing the variables to be analyzed.

parameter_list

Vector of column names to include in the analysis.

group_column

Character string specifying the grouping variable.

completed_table_statistic

Which statistic to show in the final table. Options: "min_max", "mean", "sd", "median", "min", "max".

freeform_brackets

(Optional) Add another of statistic in brackets after completed_table_statistic, eg. choose median in completed_table_statistic and choose Q1-Q3 here to get "median (Q1-Q3)" Options: "sd", "Q1_Q3",

multiple_comparison_correction

Method for correcting p-values. Options: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", NULL.

comparison_list

Optional list of specific group comparisons to make.

sex_column

Character string specifying the sex variable (optional).

female_value

Value used to identify females in sex_column.

male_value

Value used to identify males in sex_column.

completed_table_sex_statistic

Which sex statistic to show in the final table. Options: "n_m_f" n males and n females "n_fn_ratio" n with female n to n ratio, "n_f_fn_ratio", n female with n female to n ratio, "f_n_ratio", n female to n ratio "n_f", n females "n_m" n males

format_value

Number of decimal places for formatting rounded values.

Value

demographic_statistics

Raw demographic statistics for all groups

sex_ratios

Sex distribution statistics by group

complete_table

Final formatted table suitable for publication

Examples

if (FALSE) { # \dontrun{
# Using the function
results <- generate_demographic_tables()
print(results$demographic_statistics)
print(results$group_wise_comparisons)
print(results$sex_ratios)
print(results$complete_table)

} # }