run_univariate_logistic

run_univariate_logistic This function performs univariate logistic regression analyses for a set of predictor variables against a binary outcome variable in a given dataset. It returns a tidy data frame with odds ratios (OR), 95% confidence intervals (CI), and p-values for each predictor.

Usage

run_univariate_logistic(data, outcome_var, predictors)

Arguments

data

A data frame containing the outcome and predictor variables.

outcome_var

A string specifying the name of the binary outcome variable. If the variable is a factor or character, values equal to "Yes" are recoded to 1 and others to 0.

predictors

A character vector specifying the names of the predictor variables to test.

Value

A data frame with the following columns:

Predictor

Name of the predictor variable

OR

Estimated odds ratio

CI_low

Lower bound of the 95% confidence interval for the OR

CI_high

Upper bound of the 95% confidence interval for the OR

p_value

p-value for the predictor in the logistic model

Examples

# Reading the data
data<- MSNeuroLab::data
# Selecting variables to test at univariate 
variables_to_test <- colnames(data)[c(3:8)]
# Using the function
res_univariates <- run_univariate_logistic(data, "smoker", predictors = variables_to_test)
#> Waiting for profiling to be done...
#> Waiting for profiling to be done...
#> Waiting for profiling to be done...
#> Waiting for profiling to be done...
#> Waiting for profiling to be done...
#> Waiting for profiling to be done...