similarity_analysis

similarity_analysis This function performs similarity matching between patients and controls based on multiple variables. It identifies the most similar control subject for each patient using distance-based similarity metrics.

Usage

similarity_analysis(
  data = mtcars,
  input_columns = c("mpg", "cyl", "hp"),
  group_column = "am",
  patient_value = 1,
  control_value = 0,
  comparison_variable = "disp",
  distance_metric = "euclidean",
  x_label = NULL,
  y_label = NULL,
  plot_title = NULL,
  include_id = FALSE
)

Arguments

data

data.frame that contains numeric or categorical variables arranged in columns.

input_columns

List of columns that are used in computing distance metrics.

group_column

Column containing categories for the two groups whose similarity in being analysed, for example MS patient and healthy control.

patient_value

Value in group_column that identifies patients. Every patient is matched to one control. Several patients may match the same control.

control_value

Value in group_column that identifies controls. Every control may not be match to any patient.

comparison_variable

Compare this variable between patients and their most similar controls.

distance_metric

Metric used to compute difference values. Choose "euclidean", "cosine", or "manhattan".

x_label

Variable to plot on x-axis. Variable name is used if NULL.

y_label

Variable to plot on y-axis. Variable name is used if NULL.

plot_title

Plot title. Will be auto-generated if NULL.

include_id

Whether to include patient/control IDs in results table.

Value

result

List of a table with numeric comparison_variable values, distance metrics and similarities for pairs of most similar patients and controls, and a scatterplot visualizing the comparisons

Examples